Refactor nonce traits
