+++ crates/relayer-all-in-one/src/one_for_all/impls/transaction/fields.rs
@@ -1,9 +1,9 @@
 use core::time::Duration;
 
 use ibc_relayer_components::runtime::traits::mutex::HasMutex;
-use ibc_relayer_components::transaction::impls::nonces::mutex::HasMutexForNonceAllocation;
 use ibc_relayer_components::transaction::impls::poll::HasPollTimeout;
 use ibc_relayer_components::transaction::traits::fee::HasFeeForSimulation;
+use ibc_relayer_components::transaction::traits::nonce::mutex::HasMutexForNonceAllocation;
 use ibc_relayer_components::transaction::traits::signer::HasSigner;
 
 use crate::one_for_all::traits::transaction::OfaTxContext;
