+++ crates/relayer-components/src/transaction/traits/nonce/guard.rs
@@ -1,0 +1,7 @@
+use crate::transaction::traits::types::HasNonceType;
+
+pub trait HasNonceGuard: HasNonceType {
+    type NonceGuard<'a>: Send + Sync;
+
+    fn deref_nonce<'a, 'b>(guard: &'a Self::NonceGuard<'b>) -> &'a Self::Nonce;
+}
